split functions to helper file
